# backend/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "La Casa De Rozgaar — Module 2 API",
    "description": "User, Talent & Career Intelligence Backend Module for La Casa De Rozgaar platform. Handles candidate profiles, assessment engine with anti-cheat integrity signals, skill gap analysis, job matching, career simulation, learning recommendations, interview intelligence, research intelligence, employer talent discovery, and workforce intelligence.",
    "version": "1.0.0",
    "contact": {
      "name": "La Casa De Rozgaar Engineering"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api/v1",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token obtained from POST /auth/login or POST /auth/register"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/assessments": {
      "get": {
        "summary": "List available assessments",
        "tags": [
          "Assessments"
        ],
        "parameters": [
          {
            "name": "roleId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skillId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Assessments list"
          }
        }
      }
    },
    "/assessments/{id}/attempt": {
      "post": {
        "summary": "Start an assessment attempt",
        "tags": [
          "Assessments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Attempt started with questions"
          }
        }
      }
    },
    "/assessments/attempts/{attemptId}/answers": {
      "post": {
        "summary": "Submit an answer during assessment",
        "tags": [
          "Assessments"
        ],
        "parameters": [
          {
            "name": "attemptId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Answer recorded"
          }
        }
      }
    },
    "/assessments/attempts/{attemptId}/integrity": {
      "post": {
        "summary": "Report anti-cheat integrity event (tab switch, copy/paste, etc.)",
        "tags": [
          "Assessments"
        ],
        "parameters": [
          {
            "name": "attemptId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Integrity event logged"
          }
        }
      }
    },
    "/assessments/attempts/{attemptId}/submit": {
      "post": {
        "summary": "Submit completed assessment for grading and skill profile update",
        "tags": [
          "Assessments"
        ],
        "parameters": [
          {
            "name": "attemptId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Assessment graded, skill profile updated"
          }
        }
      }
    },
...
  }
}